🤖 refactor(site/src/pages/CreateTemplatePage): move drawer focus handling to parent via onCloseAutoFocus
