+++ site/src/pages/CreateTemplatePage/CreateTemplatePage.tsx
@@ -24,6 +24,9 @@ const CreateTemplatePage: FC = () => {
 	// close. The drawer is opened from buttons outside its tree, so Radix has no
 	// trigger to fall back to.
 	const buildLogsOpenerRef = useRef<HTMLElement | null>(null);
+	// Keeps focus on the variables input (rather than the opener) when the drawer
+	// closes via the "Fill variables" action.
+	const preserveVariablesFocusRef = useRef(false);
 
 	const openBuildLogs = () => {
 		buildLogsOpenerRef.current =
@@ -31,6 +34,30 @@ const CreateTemplatePage: FC = () => {
 				? document.activeElement
 				: null;
 		setIsBuildLogsOpen(true);
+	};
+
+	const fillVariables = () => {
+		variablesSectionRef.current?.scrollIntoView({ behavior: "smooth" });
+		preserveVariablesFocusRef.current = true;
+		setIsBuildLogsOpen(false);
+	};
+
+	const restoreFocusOnDrawerClose = (event: Event) => {
+		// Radix focuses a DrawerTrigger on close by default, but this drawer has
+		// none. Return focus to the variables input for the "Fill variables" flow,
+		// otherwise to whatever opened the drawer, so keyboard users are not
+		// dropped onto the document body.
+		if (preserveVariablesFocusRef.current) {
+			preserveVariablesFocusRef.current = false;
+			event.preventDefault();
+			variablesSectionRef.current?.querySelector("input")?.focus();
+			return;
+		}
+		const opener = buildLogsOpenerRef.current;
+		if (opener?.isConnected) {
+			event.preventDefault();
+			opener.focus();
+		}
 	};
 
 	const pageViewProps: CreateTemplatePageViewProps = {
@@ -72,9 +99,9 @@ const CreateTemplatePage: FC = () => {
 				error={createTemplateMutation.error}
 				open={isBuildLogsOpen}
 				onClose={() => setIsBuildLogsOpen(false)}
-				openerRef={buildLogsOpenerRef}
+				onFillVariables={fillVariables}
+				onCloseAutoFocus={restoreFocusOnDrawerClose}
 				templateVersion={templateVersion}
-				variablesSectionRef={variablesSectionRef}
 			/>
 		</>
 	);
